Total carbon placed in use in od tons of wood fiber

Corresponds to B3147 column in SW Calc P and SW Calc spreadsheet for Production and Stock Change approach respectively.

calcplacediu(Years = 1900:2020, approach = c("Production", "Stock Change"),
  total = TRUE)

Arguments

Years

years to calculate carbon placed in use for

approach

The approach type to use. Fractions dont change but inflow does

total

Whether to return total placed in use or seperate it by end uses.

Value

a data frame of carbon placed in use for various end uses

Examples

calcplacediu()

#>   [1] 56835323 58780354 60800014 63060756 64944951 66023225 70843577 75460302
#>   [9] 68943091 73124619 73457917 71140222 72901192 71594160 67151866 62894277
#>  [17] 66380690 62163488 58530215 61179440 61242168 52792088 58070730 63818729
#>  [25] 61936646 60764512 59692499 56688325 55365514 57646893 46009631 33720621
#>  [33] 25512954 28975314 31131945 35497910 41177687 43200327 37744710 42209801
#>  [41] 44407185 51111694 50941531 48141464 46970658 41524576 47886176 49658378
#>  [49] 49820904 43014887 48980319 48740350 48366793 47967365 47405877 49115079
#>  [57] 49738421 43886375 44704265 48779392 44490048 44121200 45607840 49339923
#>  [65] 52361028 55978269 56764358 55505920 57760660 58094218 57918053 61586491
#>  [73] 63981270 65214228 57572753 53718966 60141313 63808229 65819151 65769450
#>  [81] 58192967 53975296 53818077 63903729 67880795 68374294 74368236 79614022
#>  [89] 81525508 82023708 80292919 74779631 77062726 76497144 76712474 74977725
#>  [97] 75922776 77036355 76524227 78766487 78496553 74556524 76332813 74812808
#> [105] 78486222 78919662 77292861 72847000 59371388 51191395 51916199 55166562
#> [113] 56351799 59681140 61548609 64055121 66560763 69065631 71569806 74073357
#> [121] 76275030
#> 

calcplacediu(approach = "Stock Change", total = FALSE)

#>     House.SingFam House.Multifam House.MobHom Res.Upkeep New.Nonres.AllRR
#> 1        10326749        4829118         0.00    2384328        1354128.5
#> 2        10795266        5048262     26332.85    2521795        1419980.8
#> 3        11284525        5276210     55344.67    2666989        1489117.2
#> 4        11572732        5410139     85581.69    2766967        1532088.2
#> 5        11973114        5596553    118646.98    2895834        1590232.8
#> 6        12178446        5690014    152357.47    2979624        1623099.3
#> 7        12850573        6000134    195512.04    3180466        1718915.8
#> 8        12742571        5948874    227374.69    3189623        1710140.2
#> 9        11657933        5440839    239649.39    2951278        1570008.7
#> 10       12349743        5762868    287175.45    3161610        1668793.6
#> 11       12201619        5691881    317736.23    3158887        1654980.4
#> 12       11581371        5403378    331725.01    3031524        1576180.1
#> 13       12049490        5622274    376932.28    3188851        1645475.9
#> 14       11672626        5445955    397115.02    3123173        1599771.0
#> 15       10899607        5083557    402569.79    2948560        1499698.7
#> 16       10048090        4685242    400173.91    2748028        1387952.8
#> 17       10889103        5077705    463458.98    3010210        1509213.3
#> 18        9658360        4501948    440667.05    2699144        1344309.1
#> 19        8641438        4025449    422653.81    2441389        1208273.6
#> 20        9260685        4313493    480456.91    2644341        1299527.6
#> 21        9371071        4368727    505318.77    2703484        1318335.4
#> 22        7631987        3555698    436547.50    2225657        1079487.3
#> 23        9230393        4300793    554429.35    2719968        1309676.0
#> 24       10768909        5016514    681659.69    3206801        1533563.6
#> 25       10283238        4789076    683091.40    3094452        1470824.0
#> 26       10660896        4960665    751353.14    3242552        1532525.9
#> 27       10313701        4795555    766706.29    3170435        1490629.2
#> 28        9570122        4443531    758087.75    2974044        1392418.5
#> 29        9259543        4297664    766792.85    2907638        1353915.5
#> 30        9928907        4601083    867553.33    3152057        1465363.8
#> 31        7543110        3494072    687228.82    2419564        1119082.7
#> 32        5090231        2353700    494879.13    1650814         761080.0
#> 33        3437675        1582174    358412.91    1128575         525376.6
#> 34        4322102        1993714    448891.44    1431582         660256.4
#> 35        4788283        2209878    510002.52    1601489         733537.0
#> 36        5845980        2702336    631524.43    1973244         894464.5
#> 37        7119697        3289735    786196.84    2427145        1097670.5
#> 38        7430159        3432074    851758.51    2558581        1150786.1
#> 39        6424708        2959556    792246.74    2237769        1006640.2
#> 40        7463952        3433878    928284.76    2625413        1185262.0
#> 41        8072814        3713478   1032775.22    2867458        1288050.5
#> 42        9816755        4497190   1367427.81    3529534        1592925.8
#> 43        9948190        4543527   1473394.59    3617798        1638153.4
#> 44        9310654        4263747   1403611.97    3415834        1522334.8
#> 45        9073162        4150457   1432318.11    3364026        1494673.6
#> 46        7840054        3578856   1346538.33    2942241        1302255.3
#> 47        9320705        4269085   1589201.54    3525574        1536755.7
#> 48        9412608        4297996   1646583.67    3597370        1581625.6
#> 49       10106446        4608614   1868371.40    3905697        1712435.5
#> 50        8673872        3967902   1525180.76    3356839        1518668.9
#> 51       10752627        4949029   1710952.53    4153214        1924967.2
#> 52       10552794        4841355   1804320.39    4116325        1983118.3
#> 53       10579875        4868032   1772644.89    4141840        2044351.2
#> 54       10722160        4929447   1780558.25    4226871        2158147.2
#> 55       10755494        4956518   1760333.90    4263295        2225312.9
#> 56       11424624        5247628   1842864.74    4576978        2478441.1
#> 57       11685537        5388375   1827620.69    4702367        2589202.2
#> 58       10100892        4640346   1543530.00    4119090        2344447.5
#> 59       10529848        4838297   1546616.92    4333131        2524894.9
#> 60       11936202        5492213   1684298.21    4955519        2942209.2
#> 61       10795324        4972073   1441091.26    4514431        2739842.8
#> 62       10784039        4960013   1367302.28    4572239        2835293.5
#> 63       11498806        5297937   1378333.09    4926980        3099533.8
#> 64       12319087        5453938   1474282.55    5569188        3318096.5
#> 65       13009332        5521872   1553203.67    6196062        3512047.1
#> 66       14015183        5722800   1680362.38    6951038        3736761.8
#> 67       14179760        5541513   1699801.62    7354614        3777745.2
#> 68       13864463        5185705   1665983.13    7496887        3678500.5
#> 69       14729119        5236775   1765956.89    8331448        3953271.9
#> 70       14556926        4958864   1762531.19    8580554        3876838.0
#> 71       14268598        4618369   1728890.39    8686417        3793861.2
#> 72       15993524        4910506   1946499.91   10106511        4276577.4
#> 73       17212533        5011843   2110104.32   11273581        4618379.9
#> 74       17406556        4819641   2151826.97   11731389        4607443.4
#> 75       15176656        3986043   1888550.58   10490830        3948152.4
#> 76       14119159        3484676   1761890.79    9972758        3644029.0
#>     New.Nonres.Rties New.Nonres.Rcar.Repair Manu.HouseFurniture
#> 1          1103132.3               96707.94           1538427.9
#> 2          1156811.4              101413.81           1617366.0
#> 3          1212447.2              106291.22           1699755.2
#> 4          1246749.9              109298.42           1752567.8
#> 5          1293417.9              113389.65           1823039.3
#> 6          1318023.7              115546.76           1863369.6
#> 7          1392477.0              122073.83           1975125.6
#> 8          1384570.4              121380.69           1969179.8
#> 9          1269582.1              111300.05           1810853.7
#> 10         1348614.4              118228.55           1928748.6
#> 11         1335398.5              117069.95           1915407.6
#> 12         1272151.9              111525.33           1828869.3
#> 13         1328159.9              116435.37           1913903.1
#> 14         1290424.7              113127.25           1864330.2
#> 15         1207629.3              105868.85           1749803.5
#> 16         1116061.6               97841.41           1621642.1
#> 17         1213609.4              106393.10           1767564.9
#> 18         1078630.5               94559.95           1575698.9
#> 19          966435.4               84724.18           1416466.4
#> 20         1038765.3               91065.10           1526422.5
#> 21         1057317.5               92691.51           1555511.2
#> 22          862305.5               75595.46           1272793.4
#> 23         1046612.6               91753.05           1548387.5
#> 24         1224254.6              107326.33           1816208.7
#> 25         1172044.0              102749.21           1743547.5
#> 26         1215986.0              106601.45           1815673.1
#> 27         1177706.8              103245.64           1764720.9
#> 28         1091846.3               95718.53           1643614.5
#> 29         1058725.0               92814.91           1598557.9
#> 30         1133777.2               99394.48           1719969.2
#> 31          863184.6               75672.53           1313479.4
#> 32          581364.4               50966.29            889409.7
#> 33          388524.4               34060.64            599927.9
#> 34          493476.4               43261.43            760886.8
#> 35          549386.0               48162.84            848431.7
#> 36          676224.2               59282.33           1044206.3
#> 37          825361.4               72356.69           1277873.9
#> 38          863434.0               75694.39           1340846.0
#> 39          743013.3               65137.51           1162229.3
#> 40          862911.2               75648.56           1354300.8
#> 41          936105.2               82065.23           1472955.0
#> 42         1128092.6               98896.13           1792525.1
#> 43         1136499.1               99633.10           1820029.8
#> 44         1076016.8               94330.82           1720754.2
#> 45         1048674.1               91933.78           1685001.0
#> 46          903385.4               79196.80           1463107.7
#> 47         1088860.7               95456.80           1757440.9
#> 48         1093211.9               95838.25           1776123.6
#> 49         1172992.4              102832.34           1916872.0
#> 50          969258.8               88054.38           1641666.6
#> 51         1166364.8              109928.16           2032316.0
#> 52         1078235.1              105553.49           1988304.3
#> 53         1036643.5              105543.86           1994178.4
#> 54          994719.5              105475.37           2016287.8
#> 55          952623.2              105358.36           2027426.7
#> 56          948097.1              109547.48           2149710.0
#> 57          927315.5              112135.71           2207490.2
#> 58          747293.9               94756.43           1911324.7
#> 59          733750.1               97764.01           1998284.5
#> 60          784526.1              110091.77           2278212.9
#> 61          667158.6               98855.44           2060168.6
#> 62          619941.4               97269.12           2073383.5
#> 63          618897.5              103149.59           2232501.7
#> 64          651453.3              107711.89           2389412.5
#> 65          673043.8              110400.01           2527434.9
#> 66          721016.4              117310.40           2700397.5
#> 67          715282.2              115449.74           2732324.7
#> 68          688281.5              110185.74           2675469.3
#> 69          704950.7              112018.03           2867875.3
#> 70          690747.7              108747.90           2886088.3
#> 71          661520.4              103379.14           2805178.5
#> 72          718942.7              111505.56           3180708.0
#> 73          752746.9              115833.77           3472997.8
#> 74          758771.1              115670.31           3511751.5
#> 75          661913.7               99938.78           3038330.7
#> 76          602424.6               90345.96           2781448.8
#>     Manu.CommFurniture Manu.OtherProducts Shipping.Tot Other.Uses.Tot      V17
#> 1             250774.3          1275133.3      5107310        5787945 21749255
#> 2             263637.5          1340566.9      5338556        6066412 22196158
#> 3             277160.9          1408737.5      5578303        6355788 22717544
#> 4             285866.3          1452389.4      5718569        6533126 23462381
#> 5             297450.5          1510677.7      5914285        6775016 23834800
#> 6             304324.6          1543726.6      6010126        6903104 24430670
#> 7             323042.3          1635723.3      6333385        7293640 26814150
#> 8             322182.9          1630656.3      6277684        7249506 31432142
#> 9             296495.5          1499274.3      5739227        6645983 28750728
#> 10            315920.6          1596729.8      6077256        7057140 30538338
#> 11            313981.6          1585423.9      5999689        6986781 30910756
#> 12            299702.2          1513964.6      5695516        6651743 30091435
#> 13            313586.3          1584466.4      5925829        6940685 30016951
#> 14            305537.2          1543388.3      5738621        6740838 29719016
#> 15            287010.4          1448333.9      5354096        6307457 28676244
#> 16            266156.6          1342110.6      4932534        5827912 28080374
#> 17            290082.3          1462935.0      5345160        6333857 28750728
#> 18            258860.8          1303889.6      4736240        5629144 28750728
#> 19            233065.7          1171757.7      4231398        5044522 28601760
#> 20            251244.5          1262625.4      4532852        5420035 28527276
#> 21            255478.3          1287430.4      4594429        5509310 28154858
#> 22            209363.0          1053250.8      3736030        4494385 25622410
#> 23            254666.5          1281313.1      4518555        5451777 25249991
#> 24            298932.3          1502643.7      5268137        6375533 25398959
#> 25            287161.3          1442441.7      5026824        6102613 24654121
#> 26            299765.8          1501273.0      5200436        6334489 22270641
#> 27            291941.9          1458606.9      5021652        6137558 22196158
#> 28            272971.0          1357396.8      4643661        5696848 21376836
#> 29            265787.3          1319991.9      4487790        5524279 20631999
#> 30            287079.4          1419734.3      4793590        5925231 20557515
#> 31            219496.2          1084057.2      3637164        4511567 17801616
#> 32            149405.1           733224.6      2443376        3043877 14449848
#> 33            101868.2           494292.1      1631630        2046181 12364302
#> 34            128384.7           627918.8      2061762        2590718 12438786
#> 35            142984.0           700386.5      2286052        2881289 12736721
#> 36            175302.4           862465.8      2800725        3537658 13332591
#> 37            214584.2          1056211.6      3406721        4319044 14524331
#> 38            225442.0          1107956.6      3550999        4518324 15194685
#> 39            197028.4           958806.7      3048006        3900793 13705010
#> 40            229790.5          1119169.8      3529767        4537502 14375364
#> 41            250036.7          1217323.0      3814726        4921529 14375364
#> 42            307989.6          1478358.5      4586888        5963016 15343652
#> 43            315413.5          1499370.4      4608490        6032063 14896750
#> 44            296685.0          1417183.2      4339257        5687033 13705010
#> 45            291701.1          1386516.6      4212863        5549752 13481559
#> 46            255731.6          1200091.2      3613994        4794587 12587754
#> 47            304550.9          1443162.9      4332651        5748859 13258108
#> 48            309598.8          1459982.6      4339071        5796645 14002945
#> 49            336096.5          1573183.9      4635643        6231405 12662238
#> 50            286117.8          1332997.3      3912433        5357058 11098079
#> 51            347286.4          1638155.9      4804968        6676873 11470498
#> 52            347925.0          1573564.7      4557750        6513209 10874628
#> 53            348652.0          1557653.5      4488958        6544204 10427725
#> 54            356516.5          1554124.0      4433140        6619161 10055306
#> 55            360121.1          1540260.0      4367273        6656712  9757371
#> 56            390112.9          1612165.4      4505373        7048632  9384953
#> 57            400433.7          1633065.1      4543730        7244344  9012534
#> 58            356147.6          1395429.3      3821215        6252280  8640115
#> 59            376769.3          1441656.8      3904971        6532703  8342180
#> 60            434251.7          1621771.4      4354470        7436520  7969761
#> 61            393876.5          1455155.0      3860465        6742497  7597343
#> 62            405399.0          1446530.9      3784182        6759577  7299408
#> 63            442577.3          1535687.1      3983974        7253916  6926989
#> 64            499274.0          1670674.6      4226956        7533247  7671826
#> 65            561005.2          1794360.7      4417958        7707241  8044245
#> 66            614775.8          1952707.1      4716593        8035146  8342180
#> 67            656303.7          2006744.2      4720833        7863429  8416664
#> 68            675990.1          1993471.2      4575312        7438881  7671826
#> 69            789864.9          2169579.8      4782968        7631331  7671826
#> 70            848432.3          2200719.8      4729346        7332338  8938050
#> 71            859133.6          2182556.1      4549376        6904561  9712681
#> 72           1052373.5          2512921.2      5020158        7465702  8967844
#> 73           1237609.7          2784103.1      5326972        7754415  7701620
#> 74           1286744.0          2854650.8      5351618        7553682  7373891
#> 75           1126862.3          2505521.4      4636191        6324789  6375809
#> 76           1057428.8          2340054.1      4220644        5607535  5735249
#>  [ reached getOption("max.print") -- omitted 45 rows ]
#>