Calculate carbon in SWDS for the Production Method

Column names with only letters come from Calc sheet, otherwise sheet is specified (need to change names)

calculatedumpcarbonproduction(Years, totalcarbonstockchange = FALSE)

Arguments

Years

years to calculate

totalcarbonstockchange

If true, returns carbon stock change, or returns data frame of intermediate calculations

Value

Either a data frame of values from Calculation spreadsheet or total carbon output

Examples

calculatedumpcarbonproduction(1995:2020)

#>     Years       CA       CB ParamResults_V ParamResults_P       CI       CS
#> 96   1995 114.4089 36.04484      0.3930708      0.3930708 14.16818 6.233997
#> 97   1996 117.0390 36.87345      0.3673884      0.3673884 13.54688 5.960627
#> 98   1997 120.0224 37.81339      0.3822584      0.3822584 14.45449 6.359974
#> 99   1998 122.0450 38.45060      0.3898972      0.3898972 14.99178 6.596384
#> 100  1999 124.0482 39.08172      0.3902860      0.3902860 15.25305 6.711343
#> 101  2000 125.0281 39.39043      0.3743003      0.3743003 14.74385 6.487293
#> 102  2001 123.6641 38.96071      0.3447074      0.3447074 13.43004 5.909219
#> 103  2002 122.5601 38.61288      0.3494005      0.3494005 13.49136 5.936198
#> 104  2003 122.1128 38.47196      0.3231569      0.3231569 12.43248 5.470290
#> 105  2004 122.9186 38.72583      0.3156480      0.3156480 12.22373 5.378442
#> 106  2005 123.3337 38.85662      0.3079519      0.3079519 11.96597 5.265026
#> 107  2006 124.9576 39.36824      0.3244935      0.3244935 12.77474 5.620884
#> 108  2007 127.0930 40.04101      0.3203581      0.3203581 12.82746 5.644082
#> 109  2008 128.0955 40.35683      0.3162227      0.3162227 12.76175 5.615168
#> 110  2009 126.7876 39.94478      0.3120873      0.3120873 12.46626 5.485153
#> 111  2010 126.9019 39.98079      0.3079519      0.3079519 12.31216 5.417351
#> 112  2011 127.7038 40.23343      0.3079519      0.3079519 12.38996 5.451582
#> 113  2012 127.7330 40.24262      0.3079519      0.3079519 12.39279 5.452828
#> 114  2013 126.9854 40.00711      0.3079519      0.3079519 12.32026 5.420916
#> 115  2014 129.4067 40.76994      0.3079519      0.3079519 12.55518 5.524279
#> 116  2015 131.7374 41.50423      0.3079519      0.3079519 12.78130 5.623774
#> 117  2016 133.6165 42.09623      0.3079519      0.3079519 12.96361 5.703990
#> 118  2017 135.4515 42.67435      0.3079519      0.3079519 13.14165 5.782325
#> 119  2018 137.1833 43.21996      0.3079519      0.3079519 13.30967 5.856254
#> 120  2019 138.8610 43.74853      0.3079519      0.3079519 13.47244 5.927875
#> 121  2020 140.5273 44.27351      0.3079519      0.3079519 13.63411 5.999009
#>           CU       CV        CX ParamResults_Q   Dumps_S ParamResults_O
#> 96  7.934179 107.6600 2.7876789     0.01613588 0.5816151      0.7274305
#> 97  7.586252 109.9885 2.3284456     0.01596480 0.5886772      0.7083651
#> 98  8.094513 112.6958 2.7072886     0.01575570 0.5957765      0.6882236
#> 99  8.395398 115.5667 2.8709338     0.01570413 0.6038332      0.6754670
#> 100 8.541709 118.4463 2.8795910     0.01599190 0.6249912      0.6774638
#> 101 8.256555 120.9224 2.4760727     0.01611950 0.6349539      0.6727094
#> 102 7.520824 122.5833 1.6609435     0.01612050 0.6280663      0.6727515
#> 103 7.555161 124.2012 1.6179379     0.01622322 0.6264251      0.6770379
#> 104 6.962188 125.1794 0.9782032     0.01606186 0.6179313      0.6703042
#> 105 6.845290 126.0015 0.8220107     0.01603494 0.6209663      0.6691806
#> 106 6.700942 126.6482 0.6467461     0.01605844 0.6239766      0.6701613
#> 107 7.153853 127.6977 1.0494877     0.01605844 0.6321924      0.6701613
#> 108 7.183378 128.7275 1.0297855     0.01605844 0.6429960      0.6701613
#> 109 7.146577 129.6752 0.9476829     0.01605844 0.6480677      0.6701613
#> 110 6.981104 130.4217 0.7465229     0.01605844 0.6414507      0.6701613
#> 111 6.894810 131.0518 0.6301080     0.01605844 0.6420291      0.6701613
#> 112 6.938377 131.6947 0.6429408     0.01605844 0.6460860      0.6701613
#> 113 6.939963 132.3098 0.6151215     0.01605844 0.6462337      0.6701613
#> 114 6.899348 132.8581 0.5482959     0.01605844 0.6424517      0.6701613
#> 115 7.030901 133.5070 0.6488329     0.01605844 0.6547016      0.6701613
#> 116 7.157531 134.2471 0.7400843     0.01605844 0.6664931      0.6701613
#> 117 7.259623 135.0508 0.8037548     0.01605844 0.6759997      0.6701613
#> 118 7.359323 135.9131 0.8622365     0.01605844 0.6852835      0.6701613
#> 119 7.453414 136.8258 0.9126978     0.01605844 0.6940451      0.6701613
#> 120 7.544568 137.7838 0.9580534     0.01605844 0.7025331      0.6701613
#> 121 7.635102 138.7846 1.0007487     0.01605844 0.7109634      0.6701613
#>      Dumps_T    Dumps_V TotalCarbonOutput       CM       CN   Dumps_N   Dumps_O
#> 96  37.15124 -2.5397522          50210416 3.810474 58.61930 0.3674957 105.42418
#> 97  34.81485 -2.3363915          50653218 3.743328 60.54586 0.3668059 101.52599
#> 98  32.66609 -2.1487582          51117344 3.670215 62.34532 0.3653187  97.78353
#> 99  30.69131 -1.9747844          51432516 3.624396 64.04787 0.3663680  94.19295
#> 100 28.88910 -1.8022087          51997257 3.675025 65.74998 0.3771781  90.75751
#> 101 27.23577 -1.6533363          52349272 3.673938 67.40145 0.3827613  87.46592
#> 102 25.70422 -1.5315467          52340157 3.673528 69.00440 0.3827185  84.30699
#> 103 24.28986 -1.4143566          52810146 3.730131 70.61561 0.3886156  81.28108
#> 104 22.97729 -1.3125711          52987484 3.705433 72.15591 0.3860424  78.37468
#> 105 21.76925 -1.2080392          53631085 3.744153 73.68893 0.3900764  75.58933
#> 106 20.65762 -1.1116320          54016791 3.776608 75.20879 0.3934576  72.91952
#> 107 19.63972 -1.0178948          54207510 3.789942 76.69732 0.3948468  70.35868
#> 108 18.71069 -0.9290357          54099798 3.782411 78.13518 0.3940622  67.90032
#> 109 17.85834 -0.8523513          53062688 3.709901 79.46075 0.3865079  65.53382
#> 110 17.06594 -0.7923931          52321857 3.658105 80.69742 0.3811117  63.25755
#> 111 16.33550 -0.7304443          52298731 3.656489 81.89649 0.3809433  61.07289
#> 112 15.66541 -0.6700881          52528806 3.672574 83.07625 0.3826192  58.97791
#> 113 15.04740 -0.6180161          52614493 3.678565 84.22746 0.3832433  56.96799
#> 114 14.47379 -0.5736050          52920700 3.699974 85.36591 0.3854737  55.04124
#> 115 13.95594 -0.5178467          53140912 3.715370 86.48614 0.3870777  53.19371
#> 116 13.48911 -0.4668329          53449698 3.736959 87.59470 0.3893269  51.42282
#> 117 13.06723 -0.4218808          53795923 3.761165 88.69447 0.3918488  49.72574
#> 118 12.68661 -0.3806183          54178838 3.787937 89.78819 0.3946380  48.09976
#> 119 12.34358 -0.3430356          54597712 3.817223 90.87848 0.3976890  46.54226
#> 120 12.03496 -0.3086181          55051829 3.848973 91.96783 0.4009968  45.05072
#> 121 11.75804 -0.2769215          55513960 3.881283 93.05682 0.4043630  43.62255
#>           BX       CH       CK       CL       CT C_SWP_StockChange_LFDumps
#> 96  22.77507 16.56728 12.75680 285.5602 140.8577                  10.74707
#> 97  22.97592 16.27534 12.53201 298.0922 146.8184                  10.56039
#> 98  23.18644 15.95746 12.28724 310.3795 153.1783                  10.34424
#> 99  23.32940 15.75824 12.13385 322.5133 159.7747                  10.24582
#> 100 23.58557 15.97837 12.30334 334.8167 166.4861                  10.57000
#> 101 23.74524 15.97364 12.29971 347.1164 172.9734                  10.65958
#> 102 23.74110 15.97186 12.29833 359.4147 178.8826                  10.74236
#> 103 23.95429 16.21796 12.48783 371.9025 184.8188                  11.07313
#> 104 24.03473 16.11058 12.40514 384.3077 190.2891                  11.03905
#> 105 24.32666 16.27893 12.53477 396.8425 195.6675                  11.28244
#> 106 24.50161 16.42003 12.64343 409.4859 200.9325                  11.49348
#> 107 24.58812 16.47801 12.68807 422.1740 206.5534                  11.61575
#> 108 24.53926 16.44527 12.66285 434.8368 212.1975                  11.64236
#> 109 24.06884 16.13000 12.42010 447.2569 217.8127                  11.37918
#> 110 23.73280 15.90481 12.24670 459.5036 223.2978                  11.20710
#> 111 23.72231 15.89778 12.24129 471.7449 228.7152                  11.25570
#> 112 23.82667 15.96771 12.29514 484.0400 234.1667                  11.37992
#> 113 23.86554 15.99376 12.31520 496.3552 239.6196                  11.45648
#> 114 24.00443 16.08684 12.38687 508.7421 245.0405                  11.59857
#> 115 24.10432 16.15378 12.43841 521.1805 250.5648                  11.71112
#> 116 24.24438 16.24765 12.51069 533.6912 256.1885                  11.84836
#> 117 24.40143 16.35289 12.59173 546.2829 261.8925                  11.99442
#> 118 24.57511 16.46929 12.68135 558.9643 267.6749                  12.14909
#> 119 24.76511 16.59662 12.77940 571.7437 273.5311                  12.31219
#> 120 24.97110 16.73466 12.88569 584.6294 279.4590                  12.48351
#> 121 25.18072 16.87514 12.99386 597.6232 285.4580                  12.65467
#>     C_PAPER_StockChange_LFDumps TotalC_StockChange_SWDSProduction
#> 96                     6.481924                          17.22899
#> 97                     5.952681                          16.51307
#> 98                     6.918505                          17.26274
#> 99                     7.492533                          17.73836
#> 100                    7.788725                          18.35873
#> 101                    7.310030                          17.96961
#> 102                    6.038616                          16.78098
#> 103                    6.139779                          17.21291
#> 104                    5.135923                          16.17497
#> 105                    4.992414                          16.27485
#> 106                    4.800140                          16.29362
#> 107                    5.652477                          17.26823
#> 108                    5.744832                          17.38719
#> 109                    5.710500                          17.08968
#> 110                    5.439283                          16.64638
#> 111                    5.317014                          16.57271
#> 112                    5.424435                          16.80435
#> 113                    5.449933                          16.90641
#> 114                    5.395607                          16.99418
#> 115                    5.655265                          17.36638
#> 116                    5.897026                          17.74538
#> 117                    6.085864                          18.08028
#> 118                    6.263943                          18.41304
#> 119                    6.425916                          18.73810
#> 120                    6.577310                          19.06082
#> 121                    6.722836                          19.37751
#> 

calculatedumpcarbonproduction(1990:1991, totalcarbonstockchange = TRUE)

#> [1] 18.27836 18.81233
#>