Calculate carbon in SWDS for the Stock Change Method

Calculate carbon in SWDS for the Stock Change Method

calculatedumpcarbonstockchange(Years, totalcarbonstockchange = FALSE)

Arguments

Years

years to calculate

totalcarbonstockchange

Whether to return data frame of intermedate calculations or just totals

Value

If totalcarbonstockchange is false: a data frame of intermediate calculations. If totalcarbonstockchange is true: a numeric vector of stock change for total carbon from dumps.

Examples

calculatedumpcarbonstockchange(Years = 1990:2015)

#>     Years TotalCarbonOutputStockChange ParamResults_Q   Dumps_C   Dumps_D
#> 91   1990                     50789440     0.01517695 0.3496421 129.91531
#> 92   1991                     50653922     0.01638197 0.3763961 125.03896
#> 93   1992                     51521203     0.01490181 0.3482497 120.33218
#> 94   1993                     52205309     0.01503865 0.3561142 115.82271
#> 95   1994                     52891673     0.01517198 0.3639950 111.50260
#> 96   1995                     53352870     0.01613588 0.3904956 107.38209
#> 97   1996                     54077914     0.01596480 0.3916059 103.42877
#> 98   1997                     54787478     0.01575570 0.3915479  99.63476
#> 99   1998                     55587859     0.01570413 0.3959676  95.99796
#> 100  1999                     56476871     0.01599190 0.4096724  92.52093
#> 101  2000                     57095559     0.01611950 0.4174646  89.19155
#> 102  2001                     57475588     0.01612050 0.4202695  85.99909
#> 103  2002                     58306899     0.01622322 0.4290647  82.94378
#> 104  2003                     58878411     0.01606186 0.4289610  80.01154
#> 105  2004                     60241448     0.01603494 0.4381558  77.20634
#> 106  2005                     61080815     0.01605844 0.4449119  74.52072
#> 107  2006                     61324267     0.01605844 0.4466852  71.94507
#> 108  2007                     60853367     0.01605844 0.4432552  69.46997
#> 109  2008                     59250726     0.01605844 0.4315815  67.08345
#> 110  2009                     58232243     0.01605844 0.4241629  64.78602
#> 111  2010                     58163727     0.01605844 0.4236639  62.58073
#> 112  2011                     58298544     0.01605844 0.4246459  60.46530
#> 113  2012                     58507421     0.01605844 0.4261673  58.43661
#> 114  2013                     58959537     0.01605844 0.4294605  56.49286
#> 115  2014                     59492926     0.01605844 0.4333457  54.63121
#> 116  2015                     60054468     0.01605844 0.4374360  52.84854
#>            Q   Dumps_F        X        Y       AA        V        D         J
#> 91  17.13246 -5.107960 3.940466 49.98057 2.440731 13.19200 32.65350  96.76489
#> 92  18.12870 -4.876355 4.169601 52.57266 2.592087 13.95910 32.07854  97.97587
#> 93  16.49147 -4.706775 3.793037 54.72364 2.150980 12.69843 33.45131  99.94062
#> 94  16.58525 -4.509473 3.814608 56.83290 2.109259 12.77064 34.70802 102.39031
#> 95  16.67655 -4.320109 3.835608 58.90109 2.068199 12.84095 36.27801 105.44697
#> 96  17.60415 -4.120510 4.048955 61.11618 2.215081 13.55520 36.55345 107.98078
#> 97  17.37573 -3.953325 3.996417 63.21572 2.099543 13.37931 35.87794 109.39389
#> 98  17.10318 -3.794001 3.933730 65.19324 1.977518 13.16944 37.76529 111.90364
#> 99  17.03138 -3.636803 3.917218 67.09711 1.903878 13.11417 38.49159 114.36441
#> 100 17.35492 -3.477032 3.991631 69.01777 1.920659 13.36329 40.08786 117.44950
#> 101 17.42191 -3.329376 4.007039 70.89744 1.879664 13.41487 39.41433 119.28330
#> 102 17.53896 -3.192460 4.033962 72.74848 1.851044 13.50500 36.92977 118.78845
#> 103 17.90601 -3.055314 4.118382 74.62756 1.879080 13.78763 36.70399 118.24046
#> 104 17.90168 -2.932237 4.117387 76.45093 1.823372 13.78430 36.37300 117.57206
#> 105 18.28541 -2.805199 4.205643 78.30687 1.855938 14.07976 37.82401 118.16718
#> 106 18.56736 -2.685623 4.270492 80.17170 1.864830 14.29686 36.99424 117.98875
#> 107 18.64136 -2.575650 4.287513 81.99873 1.827029 14.35385 37.42673 118.18194
#> 108 18.49822 -2.475103 4.254590 83.74057 1.741839 14.24363 36.14724 117.35589
#> 109 18.01104 -2.386522 4.142540 85.32288 1.582310 13.86850 33.68380 114.85448
#> 110 17.70145 -2.297428 4.071333 86.78996 1.467081 13.63011 29.42517 109.71397
#> 111 17.68062 -2.205285 4.066542 88.20965 1.419691 13.61408 30.57783 106.68151
#> 112 17.72160 -2.115436 4.075968 89.59713 1.387483 13.64563 29.69064 103.70090
#> 113 17.78510 -2.028691 4.090572 90.95837 1.361241 13.69452 29.38314 101.20054
#> 114 17.92253 -1.943744 4.122182 92.31065 1.352274 13.80035 29.41663  99.32466
#> 115 18.08467 -1.861652 4.159474 93.65973 1.349085 13.92520 40.22888 106.12011
#> 116 18.25537 -1.782674 4.198734 95.00763 1.347900 14.05663 40.77098 111.69978
#>            K   Dumps_H  Dumps_I         R       AG    Dumps_K       AI
#> 91  30.48604 0.4626850 56.59854 16.429239 7.228865 -4.2926022 9.200374
#> 92  30.86756 0.5056715 52.67830 15.395453 6.773999 -3.9202451 8.621454
#> 93  31.48656 0.4692067 49.02825 15.243331 6.707065 -3.6500414 8.536265
#> 94  32.25834 0.4851218 45.67579 15.519709 6.828672 -3.3524598 8.691037
#> 95  33.22135 0.5040337 42.60062 14.771499 6.499460 -3.0751781 8.272040
#> 96  34.01964 0.5489366 39.80521 13.372126 5.883735 -2.7954108 7.488390
#> 97  34.46484 0.5502242 37.22764 12.661983 5.571273 -2.5775620 7.090710
#> 98  35.25554 0.5554758 34.85470 13.476726 5.929760 -2.3729411 7.546967
#> 99  36.03082 0.5658326 32.67523 14.048315 6.181258 -2.1794699 7.867056
#> 100 37.00278 0.5917449 30.68859 14.441669 6.354334 -1.9866440 8.087334
#> 101 37.58052 0.6057791 28.86887 14.066401 6.189216 -1.8197208 7.877184
#> 102 37.42462 0.6033037 27.18790 12.900542 5.676239 -1.6809651 7.224304
#> 103 37.25198 0.6043468 25.63818 13.015858 5.726978 -1.5497181 7.288881
#> 104 37.04139 0.5949538 24.19991 11.970182 5.266880 -1.4382707 6.703302
#> 105 37.22889 0.5969629 22.87497 11.751226 5.170539 -1.3249428 6.580687
#> 106 37.17267 0.5969351 21.65269 11.447394 5.036854 -1.2222775 6.410541
#> 107 37.23354 0.5979125 20.52605 12.082041 5.316098 -1.1266420 6.765943
#> 108 36.97329 0.5937333 19.48288 11.844692 5.211665 -1.0431758 6.633028
#> 109 36.18521 0.5810780 18.50888 11.442585 5.034738 -0.9739979 6.407848
#> 110 34.56568 0.5550709 17.58638 10.787510 4.746504 -0.9224987 6.041005
#> 111 33.61030 0.5397289 16.72123 10.350354 4.554156 -0.8651525 5.796198
#> 112 32.67125 0.5246492 15.90922 10.061172 4.426916 -0.8120089 5.634256
#> 113 31.88350 0.5119993 15.14847  9.818584 4.320177 -0.7607429 5.498407
#> 114 31.29250 0.5025087 14.43794  9.636585 4.240097 -0.7105357 5.396487
#> 115 33.43343 0.5368886 13.81419 10.295887 4.530190 -0.6237497 5.765697
#> 116 35.19132 0.5651176 13.26483 10.837232 4.768382 -0.5493644 6.068850
#>            AJ          AL Total_swpstock_change_LD Total_paper_stock_change_LD
#> 91   91.07684  4.84660452                 10.52477                    7.782868
#> 92   95.14982  4.07298236                 11.67483                    6.926736
#> 93   98.95568  3.80586142                 10.14263                    6.862886
#> 94  102.73562  3.77993980                 10.37043                    7.256152
#> 95  105.94323  3.20760846                 10.58904                    6.631890
#> 96  108.25661  2.31337240                 11.64977                    5.401697
#> 97  110.08490  1.82829409                 11.52553                    4.822005
#> 98  112.26522  2.18032382                 11.35296                    5.737142
#> 99  114.65156  2.38633856                 11.38124                    6.388127
#> 100 117.13926  2.48769670                 11.80691                    6.855387
#> 101 119.31290  2.17363964                 11.96516                    6.543135
#> 102 120.76428  1.45137829                 12.16359                    5.446652
#> 103 122.21107  1.44679381                 12.61139                    5.624053
#> 104 123.03299  0.82192432                 12.67543                    4.650534
#> 105 123.70040  0.66740498                 13.13050                    4.513002
#> 106 124.17497  0.47457312                 13.47607                    4.289149
#> 107 124.96708  0.79210982                 13.60523                    4.981566
#> 108 125.59620  0.62912051                 13.51036                    4.797609
#> 109 125.98171  0.38551188                 13.06429                    4.446251
#> 110 125.99953  0.01781777                 12.79977                    3.841823
#> 111 125.78290 -0.21663376                 12.82848                    3.472369
#> 112 125.42159 -0.36130413                 12.91768                    3.253603
#> 113 124.94712 -0.47447189                 13.02707                    3.084962
#> 114 124.39703 -0.55009524                 13.20888                    2.979466
#> 115 124.22439 -0.17263371                 13.41263                    3.733807
#> 116 124.34896  0.12456494                 13.62186                    4.343583
#>     Var1b_STOCKCHANGE_TOTAL
#> 91                 18.30763
#> 92                 18.60157
#> 93                 17.00552
#> 94                 17.62658
#> 95                 17.22093
#> 96                 17.05147
#> 97                 16.34753
#> 98                 17.09010
#> 99                 17.76937
#> 100                18.66230
#> 101                18.50829
#> 102                17.61024
#> 103                18.23545
#> 104                17.32596
#> 105                17.64350
#> 106                17.76522
#> 107                18.58679
#> 108                18.30797
#> 109                17.51054
#> 110                16.64159
#> 111                16.30085
#> 112                16.17128
#> 113                16.11204
#> 114                16.18834
#> 115                17.14643
#> 116                17.96544
#> 

calculatedumpcarbonstockchange(Years = 1990:2015, totalcarbonstockchange = TRUE)

#>  [1] 18.30763 18.60157 17.00552 17.62658 17.22093 17.05147 16.34753 17.09010
#>  [9] 17.76937 18.66230 18.50829 17.61024 18.23545 17.32596 17.64350 17.76522
#> [17] 18.58679 18.30797 17.51054 16.64159 16.30085 16.17128 16.11204 16.18834
#> [25] 17.14643 17.96544
#>