Estimates U.S. HWP contribution to annual greenhouse gas removals.

Uses formulas described in WOODCARB II paper to calculate contribution contribution with the HWP variables.

finalCarbonContribution(Years = 1990:2015, approach = c("Production",
  "Stock Change", "Atmospheric Flow", "Annual Harvest", "Annual CO2 Release",
  "Total Contribution"), decaydistribution = c("Exponential", "K=2", "K=10"),
  plot = FALSE, halflives = hl, paperHL = 2.53087281800454, fsp = fsp_1,
  fnsp = fnonsp, fsawn = fsw, swpdata, woodToCarbon = 4.535925e-07,
  paperToCarbon = 3.9008955e-07, swpSwdsNondegradable = 0.77,
  paperSwdsNondegradable = 0.44, swpLandfillDecay = 0.0300063714528115,
  paperLandfillDecay = 0.0478032538317204,
  swpDumpDecay = 0.0420089200339361, paperDumpDecay = 0.0840178400678722)

Arguments

Years
years to return carbon totals for

approach
The approach used to calculate carbon contribution. "Production" approach corresponds to `06 IPCC Tables`$R9 in the WOODCARB II spreadsheet "Stock Change" approach corresponds to `06 IPCC Tables`$P9 in the WOODCARB II spreadsheet "Atmospheric Flow" approach corresponds to `06 IPCC Tables`$Q9 in the WOODCARB II spreadsheet

decaydistribution
Type of decay method to use

plot
if true, returns a simple plot of time vs. carbon contribution.

halflives
data frame of half lives to use. Must have a column associated with each one of 13 end uses and half life values for 1900 until the maximum year of interest.

paperHL
half life value for paper

fsp
Fraction of structural panel products that go to each end use. Default can be substituted with data with a column for each end use and row for years from 1900 to latest year of interest.

fnsp
Fraction of non-structural panel products that go to each end use. Default can be substituted with data with a column for each end use and row for years from 1900 to latest year of interest.

fsawn
Fraction of sawnwood products that go to each end use. Default can be substituted with data with a column for each end use and row for years from 1900 to latest year of interest.

swpdata
Optional argument to provide swp production and trade time series for purposes of error analysis

woodToCarbon
Conversion factor to convert oven-dry tons of solid wood to Tg of Carbon.

paperToCarbon
Conversion factor to convert oven-dry tons of paper to Tg of Carbon.

swpSwdsNondegradable
Percent of solid-wood products in landfills and dumps that are not subject to decay.

paperSwdsNondegradable
Percent of paper products in landfills and dumps that are not subject to decay.

swpLandfillDecay
Decay rate of solid-wood products in landfills as a half-life in years.

paperLandfillDecay
Decay rate of paper products in landfills as a half-life in years.

swpDumpDecay
Decay rate of solid-wood products in dumps as a half-life in years.

paperDumpDecay
Decay rate of paper products in dumps as a half-life in years.

Value

A vector of carbon contributions for all Years. Plot returned is optional All units returned are in Thousand metric tons CO2 Equivalent.

Details

Decay type and half life options allow for analysis of contribution levels under different circumstances.

The Production Approach focuses on estimating annual carbon stock change in HWP and forests where the carbon is from trees harvested in the reporting country.

The Stock Change Approach focuses on estimating annual carbon stock change in HWP and forests in a country regardless of wood origin.

The Atmospheric Flow Approach focuses on estimating annual carbon fluxes between the atmosphere and forests/HWPs within a country.

Examples

finalCarbonContribution()

#>  [1] -131771.55 -123758.01 -123790.61 -120707.52 -122498.31 -118411.13
#>  [7] -112219.10 -117343.83 -114188.45 -119182.37 -112968.80  -93478.73
#> [13]  -98188.13  -93967.40 -103967.46 -102683.17 -107666.14 -102762.70
#> [19]  -76830.26  -54360.98  -60549.58  -68943.22  -68312.97  -70814.78
#> [25]  -86538.26  -91250.87

finalCarbonContribution(approach = "Stock Change")

#>  [1] -129621.60 -116344.77 -119985.28 -126805.46 -129953.63 -125980.54
#>  [7] -122340.29 -131434.35 -137217.83 -147057.27 -141195.29 -125039.45
#> [13] -130714.01 -125812.07 -143193.24 -142102.08 -138121.66 -115131.55
#> [19]  -73079.21  -42305.07  -49151.35  -52374.83  -57982.74  -67784.59
#> [25] -110330.38 -115278.27

finalCarbonContribution(approach = "Production",
                         decaydistribution ="Exponential")

#>  [1] -131771.55 -123758.01 -123790.61 -120707.52 -122498.31 -118411.13
#>  [7] -112219.10 -117343.83 -114188.45 -119182.37 -112968.80  -93478.73
#> [13]  -98188.13  -93967.40 -103967.46 -102683.17 -107666.14 -102762.70
#> [19]  -76830.26  -54360.98  -60549.58  -68943.22  -68312.97  -70814.78
#> [25]  -86538.26  -91250.87