Values for final 7 variables.

Values for final 7 variables. All units are in Gg Carbon Var1A = Annual Change in stock of HWP in use form consumption Var1B = Annual Change in stock of HWP in SWDS from consumption Var2A = Annual Change in stock of HWP in use produced from domestic harvest Var2B = Annual Change in stock of HWP in SWDS produced from domestic harvest Var3 = Annual Imports of wood, paper products, wood fuel, pulp, recovered paper, roundwood/chips Var4 = Annual Exports of wood, paper products, wood fuel, pulp, recovered paper, roundwood/chips Var5 = Annual Domestic Harvest Var6 = Annual release of carbon to the atmosphere from HWP consumption (from fuelwood & products in use and products in SWDS) Var 7 = Annual release of carbon to the atmosphere from HWP (including fuelwood) where wood came from domestic harvest (from products in use and products in SWDS)

finalVariables(Years = 1990:2015, Variable = paste("Var", c("1A", "1B",
  "2A", "2B", "3", "4", "5", "6", "7"), sep = ""),
  decaydistribution = c("Exponential", "K=2"), halflives = halfLives)

Arguments

Years

years to calculate

Variable

variable to return

decaydistribution

decay distribution to use (if necessary)

halflives

half-lives for calculation (if necessary)

Value

a numeric vector of values for the selected variable for `Years`

Examples

finalVariables(Variable = "Var1A")

#>  [1] 17043.7116 13128.8249 15717.7392 16956.7233 18220.9725 17306.8633
#>  [7] 17018.0011 18755.6290 19653.6773 21444.2277 19999.5154 16491.4294
#> [13] 17413.8298 16986.4185 21409.1983 20989.8919 19082.7521 13091.5426
#> [19]  2420.1492 -5103.8430 -2895.9387 -1887.2390  -298.5597  2298.3609
#> [25] 12943.6684 13474.0869
#> 

finalVariables(Variable = "Var2B")

#>  [1] 18278.36 18812.33 17426.64 17949.01 17479.06 17228.99 16513.07 17262.74
#>  [9] 17738.36 18358.73 17969.61 16780.98 17212.91 16174.97 16274.85 16293.62
#> [17] 17268.23 17387.19 17089.68 16646.38 16572.71 16804.35 16906.41 16994.18
#> [25] 17366.38 17745.38
#>