Paper and board production and fibrous materials consumed in the manufacture of paper and board, 1965-1999

Dataset containing consumption and production statistics for various board and paper products. Units are specified below.

howard46

Format

A data frame with 56 observations of 19 variables

Source

American Paper Institute, American Forest & Paper Association: Statistics of Paper, Paperboard & Wood Pulp

Details

Prod.PaperBoardProd.PaperBoard. Paper and Paperboard Production, in 1000 Tons

Consump.Tot Total Consumption of Fibrous Materials, in 1000 Tons

Consump.WoodPulp Consumption of Woodpulp, in 1000 Tons

Consump.RecPaper Consumption of Recovered Paper, in 1000 Tons

Consump.Other Consumption of Other Fibrous Materials, in 1000 Tons

Consump.PerTonProd.Tot Total Consumption of Fibrous Material per ton of Paper and Board Produced, in Tons

Consump.PerTonProd.WoodPulp Consumption of Wood pulp per ton of Paper and Board Produced, in Tons

Consump.PerTonProd.RecPaper Consumption of Recovered Paper per ton of Paper and Board Produced, in Tons

Consump.PerTonProd.Other Consumption of Other Fibrous Materials per ton of Paper and Board Produced, in Tons

RagsOther.RecPaperUtiRatePerc Rags and Other: Recovered Paper Utilization Rate Percent

RagsOther.Prod.Estimated Rags and Other: Estimated Production, in Thousand Tons

RagsOther.Imports.Estimated Rags and Other: Estimated Imports, in Thousand Tons

RagsOther.Exports.Estimated Rags and Other: Estimated Exports, in Thousand Tons