Wood pulp production, imports, exports, and consumption, 1965-1999

Dataset containing woodpulp statistics. All units are in thousand pounds unless otherwise specified. Consumption = Production + Imports - Exports

howard49

Format

A data frame with 56 observations of 15 variables

Source

U.S. Department of Commerce, Bureau of the Census (52,56,57); American Forest and Paper Association (2,4); United Nations, Food and Agriculture Organization (29). 1993-1997 revised to include dissolving and special alpha pulps, which is consistent with earlier years. American Paper Institute; American Forest & Paper Association: Statistics of Paper, Paperboard & Wood Pulp; 1977, 1982, 1990, 1995, 1998, 1999 annual Statistics reports (4). 1965-1967, 1970, 1973, 1977-1978, 1986-1988 & 1997 revised. 1966, 1968, 1972, 1975, 1985-1987 & 1996-1997 revised.

Details

Production Wood pulp Production

Imports Wood pulp Imports

Imports.PercConsump Imports as Percent of Consumption

Exports Wood pulp Exports

Exports.PercProd Exports as Percent of Production

Consump.Tot Total Consumption

Consump.PerCap(pounds) Per Capita Consumption in pounds