Production, imports, exports, and consumption of timber products, by major product, 1965-1999

Dataset containing statistics for various timber products. Prior to 1989, pulpwood chips are not included in total production. All units are in million cubic ft.

howard5

Format

A data frame with 56 observations of 24 variables

Source

U.S. Department of Agriculture, Forest Service(16, 69); Data may not add to totals because of rounding; Data have been revised.

Details

AllProducts.Prod Production for all timber products

AllProducts.Consump Consumption for all timber products

Prod.Tot Total Production for Industrial Roundwood

Imports.Tot Total Imports for Industrial Roundwood

Exports.Tot Total Exports for Industrial Roundwood

Consump.Tot Total Consumption for Industrial Roundwood

Lumber.Prod Production for Lumber Products

Lumber.Imports Imports for Lumber Products

Lumber.Exports Exports for Lumber Products

Lumber.Consump Consumption for Lumber Products

PlywoodVeneer.Prod Production for Plywood and Veneer Products

PlywoodVeneer.Imports Imports for Plywood and Veneer Products

PlywoodVeneer.Exports Exports for Plywood and Veneer Products

PlywoodVeneer.Consump Consumption for Plywood and Veneer Products

Pulpwood.Prod Production for Pulpwood Products

Pulpwood.Imports Imports for Pulpwood Products

Pulpwood.Exports Exports for Pulpwood Products

Pulpwood.Consump Consumption for Pulpwood Products

OtherIndProductsandConsump Other Industrial Products, production and consumption.

Logs.Imports Log Imports

Logs.Exports Log Exports

PulpwoodChip.Imports Pulpwood chip Imports

PulpwoodChip.Exports Pulpwood chip Exports

FuelwoodProdandConsump Fuelwood production and consumption