Imports and exports of veneer, by kind, in the US 1935-1962 (estimates of 1927-34).

Dataset containing import and export statistics for veneer products. All units are in 1000 Sq. Ft. All vectors are numeric.

ie_veneer

Format

A data frame with 36 observations on 6 variables:

Years

Years

Imports.Tot

Total Veneer Imports

Imports.BirchMaple

Birch or Maple Veneer Imports

Imports.Other

Other Products Imports

Exports.Tot

Total Veneer Exports

Exports.Fancy.Face.Figured.Special

Exports of fancy, face, figured and special Veneer Products

Exports.UtilityCommercialContainer

Exports of utility, commerical and container Veneer Products

Source

US Department of Commerce, Bureau of the Census