U.S. Annual industrial wood product in conventional units of measure.

U.S. Annual industrial wood product in conventional units of measure. :(units are in Million feet squared 3/8in.):

ince1

Format

A data frame with 151 observations of 25 variables

SW.Ply (Softwood Plywood units: million ft^2, 3/8 in)

OSB.Wafer.board (OSB/ Wafer board units: million ft^2, 3/8 in)

Lam.Ven.Lumb(Laminated Veneer lumber units: million ft^3)

HW.Ply.Ven(Hardwood, Plywood and veneer units: million ft^2, 3/8 in)

SW.Lumb(Softwood lumber units: million bd. ft.)

HW.Lumb(Hardwood lumber units: million bd. ft.)

Lumb.Pallet.Plant(Lumber made at Pallet plants units: million bd. ft.)

PartBoard.Prod(Particalboard production units: million ft^2, 3/4 in.)

Hardboard.Prod(Hardboard production units: million ft^2, 1/8 in.)

MDF.Pro(MDF production units: million ft^2, 3/4 in.)

Pulp.Paper.Board(Pulp, paper and board production units: Thousand tons)

Other.Products(Other industrial products units: million ft^3)

InsulatingBoard(Insulating board units: million ft^2, 1/2 in.)

Fuelwood(Fuelwood units: million ft^3)

FuelwoodT(Fuelwood units: Thousand short tons)

Log.Chip.Exports(Log and Chip exports units: Thousand short tons)

RW.Dom.Prod(Roundwood Equivalent of Domestic Product Output units: Thousand short tons)

Estimated.Tot.Dom.Timber(Estimated Total Domestic Timber Harvest units: Thousand short tons)

Source

Sources listed on the excel spread from the National Forest Services