Total production, productivity, recovered paper utilization, and per capita production

Dataset containing industrial wood and roundwood statistics, productivity, recovered paper utilization rate and per capita indudustrial wood production.

pt_pu

Format

A data frame with 151 observations of 15 variables:

Years

Years

Total.Industrial.Wood.Product.Production

Industrial wood production in thousand tons

Roundwood.Equivalents.of.Production.Hardwoods

Hardwood production in million ft<sup>3</sup>

Roundwood.Equivalents.of.Production.Softwoods

Softwood production in million ft<sup>3</sup>

Roundwood.Equivalents.of.Production.Totals.millionftcubed

Production totals in million ft<sup>3</sup>

Roundwood.Equivalents.of.Production.Totals.thousand.short.tons

Production totals in thousand short tons

Roundwood.Equivalents.of.Production.Totals.thousand.metric.tons

Production totals in thousand metric tons

Industrial.Wood.Productivity.lbs.ftsquared

Industrial wood product output per unit of roundwood input in (lbs/ft3)

Industrial.Wood.Productivity

Industrial wood product output per unit of roundwood input in tons/tons

Recovered.Paper.Utilization.Rate(AF&PA)

Recovered paper utilization rate

U.S.Population

United States population in millions

Per.Capita.Industrial.Wood.Product.Production

Per capita industrial wood product production, lbs/capita

HW.Agrifiber

Hardwood Agrifiber (units?)

HW.Roundwood

Hardwood Roundwood (units?)

U.S.Timber.Harvest

United States Timber Harvest

Misc

This is unnamed column in Ince_Table3

Source

Ince tables