Summary of Total Wood Pulp Statistics in the United States from 1899 to 1972 with necessary statistics from 1869, 1879, 1889.

New Supply is Production plus Imports minus Exports All units are in tons of 2000 pounds, air dry weight. All vectors are numeric.

twp

Format

A data frame with 77 observations of 11 variables:

Years

Years

Woodpulp.Prod

Production of Wood Pulp

Woodpulp.Imports

Imports of Wood Pulp

Woodpulp.Exports

Exports of Wood Pulp

Woodpulp.NewSupply

New Supply of Wood Pulp

Consump.Paper.Board

Consumption in Paper & Board

WastePaper.Estimated.Prod

Estimated Production of Waste Paper

WastePaper.Estimated.Imports

Estimated Imports of Waste Paper

WastePaper.Estimated.Exports

Estimated Exports of Waste Paper

Rags.Estimated.Prod

Estimated Production of Rags and Other

Rags.Estimated.Imports

Estimated Imports of Rags and Other

Rags.Estimated.Exports

Estimated Exports of Rags and Other

Source

1869 - 1962 U. S. Bureau of the Census 1963 - Pulp Producers Division, A. P. I. Imports, Exports, Consumption - Bureau of the Census