Production, imports, exports and consumption of insulating board, 1950-1987

Dataset contains statistics for insulating board. All units are in million square ft unless otherwise specified. All vectors are numeric.

ulrich54

Format

A data frame with 72 observations of 5 variables

Source

U.S. Department of Commerce, Bureau of the Census (48,51, 52); American Paper Institute (3, 4); American Hardboard Association. Data from Ince,2000, Industrial Wood Productivity in the United States, 1900-1998.

Details

HardBoard.Prod HardBoard Production

HardBoard.Import HardBoard Imports

HardBoard.Exports HardBoard Exports

HardBoard.Consump.Tot Total HardBoard Consumption

HardBoard.Consump.PerCapita HardBoard Consumption Per Capita